+++ fractal_server/app/routes/api/v2/__init__.py
@@ -16,6 +16,7 @@ from .task_collection import router as task_collection_router_v2
 from .task_collection_custom import router as task_collection_router_v2_custom
 from .task_group import router as task_group_router_v2
 from .task_group_lifecycle import router as task_group_lifecycle_router_v2
+from .task_version_update import router as task_version_update_router_v2
 from .workflow import router as workflow_router_v2
 from .workflow_import import router as workflow_import_router_v2
 from .workflowtask import router as workflowtask_router_v2
@@ -55,6 +56,7 @@ router_api_v2.include_router(
 )
 
 router_api_v2.include_router(task_router_v2, prefix="/task", tags=["V2 Task"])
+router_api_v2.include_router(task_version_update_router_v2, tags=["V2 Task"])
 router_api_v2.include_router(
     task_group_router_v2, prefix="/task-group", tags=["V2 TaskGroup"]
 )
